[utils] Remove read/write memory barriers.
For simplicity's sake, we now only expose a full read/write barrier.
