+++ mono/metadata/loader.c
@@ -41,7 +41,7 @@
 #include <mono/metadata/verify-internals.h>
 #include <mono/utils/mono-logger-internal.h>
 #include <mono/utils/mono-dl.h>
-#include <mono/utils/mono-membar.h>
+#include <mono/utils/atomic.h>
 #include <mono/utils/mono-counters.h>
 #include <mono/utils/mono-error-internals.h>
 #include <mono/utils/mono-tls.h>
