+++ mono/metadata/sgen-protocol.c
@@ -29,7 +29,7 @@
 #include "sgen-memory-governor.h"
 #include "sgen-thread-pool.h"
 #include "sgen-client.h"
-#include "mono/utils/mono-membar.h"
+#include "mono/utils/atomic.h"
 
 #include <errno.h>
 #include <string.h>
