+++ mono/mini/mini-trampolines.c
@@ -12,7 +12,7 @@
 #include <mono/metadata/tabledefs.h>
 #include <mono/utils/mono-counters.h>
 #include <mono/utils/mono-error-internals.h>
-#include <mono/utils/mono-membar.h>
+#include <mono/utils/atomic.h>
 #include <mono/utils/mono-compiler.h>
 
 #include "mini.h"
