+++ mono/utils/hazard-pointer.c
@@ -9,7 +9,6 @@
 #include <string.h>
 
 #include <mono/utils/hazard-pointer.h>
-#include <mono/utils/mono-membar.h>
 #include <mono/utils/mono-memory-model.h>
 #include <mono/utils/monobitset.h>
 #include <mono/utils/lock-free-array-queue.h>
@@ -130,7 +129,7 @@ mono_thread_small_id_alloc (void)
 
 	if (id > highest_small_id) {
 		highest_small_id = id;
-		mono_memory_write_barrier ();
+		mono_memory_barrier ();
 	}
 
 	mono_mutex_unlock (&small_id_mutex);
@@ -249,7 +248,7 @@ mono_hazard_pointer_save_for_signal_handler (void)
 		g_assert (!hp_overflow->hazard_pointers [i]);
 	*hp_overflow = *hp;
 
-	mono_memory_write_barrier ();
+	mono_memory_barrier ();
 
 	memset (hp, 0, sizeof (MonoThreadHazardPointers));
 
@@ -276,11 +275,11 @@ mono_hazard_pointer_restore_for_signal_handler (int small_id)
 
 	*hp = *hp_overflow;
 
-	mono_memory_write_barrier ();
+	mono_memory_barrier ();
 
 	memset (hp_overflow, 0, sizeof (MonoThreadHazardPointers));
 
-	mono_memory_write_barrier ();
+	mono_memory_barrier ();
 
 	overflow_busy [small_id] = 0;
 }
