+++ mono/utils/mono-linked-list-set.h
@@ -11,7 +11,7 @@
 #define __MONO_SPLIT_ORDERED_LIST_H__
 
 #include <mono/utils/hazard-pointer.h>
-#include <mono/utils/mono-membar.h>
+#include <mono/utils/atomic.h>
 
 typedef struct _MonoLinkedListSetNode MonoLinkedListSetNode;
